Read Description
Docker-compose files edited to clearer define the build - having the build just ./worker.Dockerfile was the reason as to why I couldn't get the docker build to work on my local machine but this updated one fixed that issue. Also did volumes mean to have a hyphen?

Created myself as a user in the users.cfg which works.

Added fetch_resolution that works from the available_resolutions.json. Guessed this might not be the preferred way but it works. This function used to be run from the index.html javascript whenever there is a change on the platform selection, but will scrap that idea because I guess that won't work with the actual front-end.